Bumping the version.
Fixed charting so we actually see graphs that make sense. If multiple entries are provided for the same day, it would previously create a bar for each one and repeat the date on the abscissa. Now, data is grouped by date and hours are summed for the plot. You can still see all the raw data by exporting.